+++ hedera-node/hedera-token-service-impl/src/main/java/module-info.java
@@ -27,7 +27,7 @@ module com.hedera.node.app.service.token.impl {
             com.hedera.node.app.service.token.impl.test;
     exports com.hedera.node.app.service.token.impl.serdes;
     exports com.hedera.node.app.service.token.impl;
-    exports com.hedera.node.app.service.token.impl.records to
+    exports com.hedera.node.app.service.token.records.records to
             com.hedera.node.app;
     exports com.hedera.node.app.service.token.impl.api to
             com.hedera.node.app,
