+++ hedera-node/hedera-token-service/src/main/java/com/hedera/node/app/service/token/records/TokenMintRecordBuilder.java
@@ -14,33 +14,24 @@
  * limitations under the License.
  */
 
-package com.hedera.node.app.service.token.impl.records;
+package com.hedera.node.app.service.token.records;
 
-import com.hedera.hapi.node.base.AccountID;
-import com.hedera.pbj.runtime.io.buffer.Bytes;
 import edu.umd.cs.findbugs.annotations.NonNull;
+import java.util.List;
 
 /**
  * A {@code RecordBuilder} specialization for tracking the side effects of a {@code CryptoCreate}
  * transaction.
  */
-public interface CryptoCreateRecordBuilder {
+public interface TokenMintRecordBuilder {
     /**
      * Tracks creation of a new account by number. Even if someday we support creating multiple
      * accounts within a smart contract call, we will still only need to track one created account
      * per child record.
      *
-     * @param accountID the {@link AccountID} of the new account
+     * @param serialNumbers the list of new serial numbers minted
      * @return this builder
      */
     @NonNull
-    CryptoCreateRecordBuilder accountID(@NonNull AccountID accountID);
-
-    /**
-     * The new EVM address of the account created by this transaction.
-     * @param evmAddress the new EVM address
-     * @return this builder
-     */
-    @NonNull
-    CryptoCreateRecordBuilder evmAddress(@NonNull final Bytes evmAddress);
+    TokenMintRecordBuilder serialNumbers(@NonNull List<Long> serialNumbers);
 }
